requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.31.1
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.0.2
scikit-learn==1.5.2
polars==1.8.2
openpyxl==3.1.5
python-multipart==0.0.9
Pillow==10.0.0  # For image processing and overlay generation
PyMuPDF==1.23.0  # For PDF processing and overlay rendering
# Optional (enable when you’re ready):
docling==2.*
faiss-cpu==1.8.0
mineru==2.*       # if you plan to integrate it
boto3==1.35.36
botocore==1.35.36
openai==1.3.0    # For OpenAI API support
requests==2.32.3
urllib3==2.2.3
aiohttp
tiktoken==0.5.0  # For tokenization in context extraction